compiler warnings fixed

+++ str2num.c
@@ -97,12 +97,12 @@ static long double powil (uint8_t base, uint32_t expo)
 
          if (p > 0xffffffff)
          {
-            long double fp = p;
+            long double fp = (long double) p;
 
             if (expo & 1)
                val = fp * v;
             else
-               val = v;
+               val = (long double) v;
 
             while(expo >>= 1)
             {
@@ -118,7 +118,7 @@ static long double powil (uint8_t base, uint32_t expo)
             v *= p;
       }
 
-      val = v;
+      val = (long double) v;
    }
 
    return (val);
@@ -148,12 +148,12 @@ static double powi (uint8_t base, uint32_t expo)
 
          if (p > 0xffffffff)
          {
-            double fp = p;
+            double fp = (double) p;
 
             if (expo & 1)
                val = fp * v;
             else
-               val = v;
+               val = (double) v;
 
             while(expo >>= 1)
             {
@@ -169,7 +169,7 @@ static double powi (uint8_t base, uint32_t expo)
             v *= p;
       }
 
-      val = v;
+      val = (double) v;
    }
 
    return (val);
